Better command line parsing with getopt
Removed the old command ine system and replaced
it with a getopt implementation. Previously commands
had to be defined in a specific order and command
line errors were asy to run into. getopt provides
an easy way to parse the command line and makes
specifying options straightforward and easy. It
also make the code much more readable.

Change-Id: I70b766f1711e42a6210fdd39231938cbef969100
Reviewed-on: http://review.membase.org/4501
Tested-by: Michael Wiederhold <mike@membase.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>